requirements.txt
Django==1.8
django-celery
redis
requests
